[002] - Add UserInput and ResultsTable components
