+++ src/components/ResultsTable/ResultsTable.js
@@ -1,0 +1,26 @@
+const ResultsTable = () => {
+    return (
+        <table className="result">
+            <thead>
+            <tr>
+                <th>Year</th>
+                <th>Total Savings</th>
+                <th>Interest (Year)</th>
+                <th>Total Interest</th>
+                <th>Invested Capital</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr>
+                <td>YEAR NUMBER</td>
+                <td>TOTAL SAVINGS END OF YEAR</td>
+                <td>INTEREST GAINED IN YEAR</td>
+                <td>TOTAL INTEREST GAINED</td>
+                <td>TOTAL INVESTED CAPITAL</td>
+            </tr>
+            </tbody>
+        </table>
+    );
+};
+
+export default ResultsTable;
